requirements.txt
# ─── core app ─────────────────────────────────────────
streamlit==1.33.0                
llama-cpp-python==0.2.76         # required by FlashRank 0.2.6
flashrank==0.2.6

# ─── LangChain stack ─────────── 
langchain==0.2.3
langchain-community==0.2.4
langchain-core==0.2.5
langchain-text-splitters==0.2.1   # RecursiveCharacterTextSplitter

# ─── vector store / embeddings ───────────────────────
sentence-transformers==3.0.0      # pulls torch/transformers automatically
faiss-cpu==1.9.0.post1            # works on Linux; builds if wheel absent
huggingface-hub==0.32.4
tqdm==4.67.1

# ─── PDF & misc ───────────────────────────────────────
PyPDF2==3.0.1

